Make NodeId and SymbolId into uint64s, fix up KeyBuilder (microsoft/typescript-go#243)

+++ tsc/internal/ast/ids.go
@@ -1,6 +1,6 @@
 package ast
 
 type (
-	NodeId   uint32
-	SymbolId uint32
+	NodeId   uint64
+	SymbolId uint64
 )
